Add a yml file for model configuration, update training and testing scripts accordingly
