+++ vbfml/config/model_params.yml
@@ -1,0 +1,37 @@
+conv:
+  batch_buffer_size_train: 100
+  batch_buffer_size_val: 10
+  batch_size_train: 10
+  batch_size_val: 100
+  dropout: 0.4
+  features:
+  - EventImage_pixelsAfterPUPPI
+  train_size: 0.67
+  weight_expression: Normalization
+dense:
+  batch_buffer_size_train: 1000000.0
+  batch_buffer_size_val: 10
+  batch_size_train: 20
+  batch_size_val: 1000000.0
+  dropout: 0.5
+  features:
+  - mjj
+  - dphijj
+  - detajj
+  - mjj_maxmjj
+  - dphijj_maxmjj
+  - detajj_maxmjj
+  - recoil_pt
+  - dphi_ak40_met
+  - dphi_ak41_met
+  - ht
+  - leadak4_pt
+  - leadak4_eta
+  - trailak4_pt
+  - trailak4_eta
+  - leadak4_mjjmax_pt
+  - leadak4_mjjmax_eta
+  - trailak4_mjjmax_pt
+  - trailak4_mjjmax_eta
+  train_size: 0.5
+  weight_expression: weight_total*xs/sumw
